+++ src/Vaisseau/Vaisseau.cpp
@@ -20,15 +20,15 @@ Vaisseau::Vaisseau() ///constructeur
 
 Vaisseau::~Vaisseau() ///destructeur
 {
-	for (int i = 0; i < capacites_.size(); i++)
+	for(unsigned int i = 0; i < capacites_.size(); i++)
 		delete capacites_[i];
 }
 
 void Vaisseau::gestionCapacite(std::vector <Projectile*> &projectiles, sf::Time t_ecoule)
 {
 	// Gestion des capacites_
-	for (int i = 0; i < capacites_.size(); i++)
-		if (capacites_[i] != nullptr) capacites_[i]->actualiser(projectiles, *this, t_ecoule.asMilliseconds());
+	for(unsigned int i = 0; i < capacites_.size(); i++)
+		if (capacites_[i] != nullptr) capacites_[i]->actualiser(projectiles, *this, (float)t_ecoule.asMilliseconds());
 }
